Adds support for longer guard bar to EAN13 and EAN8
  Adds a parameter to both classes that controls if we want longer guard
bars and then while building the code to be generated, use "G" instead of
"1" in those places we will need to have a longer bar.

  During the write, interpret every "G" the same as "1" but add an height
factor to the mod list. To avoid changing the entire code base to add a
height parameter, set the self.module_height for every bar generated.