Make ImageWriter font size closer to SVGWriter font size
  For some reason, that seems to be undocumented in the git history, the
ImageWriter font_size was multiplied by 2. The dpi was not taken into
account and if we changed the dpi the scale of the text changed in relation
to the bar code.

  The best match seems to be converting the font_size in pt to px. The
fractional part of the size is truncated because the library expects an
integer but it could round to the nearest integer value if this causes
problems.