Make ImageWriter and SVGWriter create similar bar codes without arguments
  If the user tries to create the same bar code with both writers, the
result should look the same without a need to tweak any parameter. It is
impossible to make both have the same *size*, as SVG is a vectorial format,
but they should have the same proportions.

  The BaseWriter should calculate all values in mm. The conversion to px is
done entirely in *Writer classes. The SVGWriter class doesn't need to
convert to px because all its values are already in mm, so its dpi
attribute is not necessary.

  Remove tweaks from the test_manually.py file.