Fix deprecation warning about ImageFont.getsize
python-barcode/barcode/writer.py:441: DeprecationWarning: getsize is
deprecated and will be removed in Pillow 10 (2023-07-01). Use getbbox
or getlength instead.
    width, height = font.getsize(subtext)

`ImageFont.getsize` is deprecated since the version 9.2 of Pillow. We
use anchor instead of computing the right position to have an aligned
text.

It is related to the issue #163.
The solution was suggested by nulano (contributor of Pillow).